requirements.txt
requests
cryptography
pyjwt
httplib2